pyproject.toml
[tool.poetry]
name = "auth_sprint_1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9.6"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
